refactor(web): dedup extractBytes into shared utils/bytes
The identical extractBytes helper was copy-pasted in acl and forward.
Move it next to getBytes in the shared bytes util and drop the two
per-module copies.
